14.03.v1

+++ Proyectos/Semana1/Secuenciales1/src/Main3.java
@@ -1,14 +1,14 @@
 import java.util.Scanner;
 
-public class Main3 {
+public class Main3 { // Main3 identifica a la clase
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         double pagoPorHora;
         double nuevoSueldo;
         double sueldo;
-        int horas;
+        long horas;
         System.out.println("Ingrese pago por hora:");
-        pagoPorHora = sc.nextDouble();
+        pagoPorHora = sc.nextDouble(); // el nombre de la variable se le denomina identificado
         System.out.println("Ingrese horas trabajadas:");
         horas = sc.nextInt();
         sueldo = horas*pagoPorHora;
